+++ tests/reference_exports/armature/armature_with_physics.escn
@@ -3,7 +3,7 @@
 [sub_resource id=1 type="ConvexPolygonShape"]
 
 resource_name = "Cube002"
-points = PoolVector3Array(-0.8, -0.8, 1.2, -0.8, 0.8, -1.2, -0.8, -0.8, -1.2, -0.8, 0.8, 1.2, 0.8, 0.8, -1.2, -0.8, 0.8, -1.2, 0.8, 0.8, 1.2, 0.8, -0.8, -1.2, 0.8, 0.8, -1.2, 0.8, -0.8, 1.2, -0.8, -0.8, -1.2, 0.8, -0.8, -1.2, 0.8, 0.8, -1.2, -0.8, -0.8, -1.2, -0.8, 0.8, -1.2, -0.8, 0.8, 1.2, 0.8, -0.8, 1.2, 0.8, 0.8, 1.2, -0.8, -0.8, 1.2, -0.8, 0.8, 1.2, -0.8, 0.8, -1.2, -0.8, 0.8, 1.2, 0.8, 0.8, 1.2, 0.8, 0.8, -1.2, 0.8, 0.8, 1.2, 0.8, -0.8, 1.2, 0.8, -0.8, -1.2, 0.8, -0.8, 1.2, -0.8, -0.8, 1.2, -0.8, -0.8, -1.2, 0.8, 0.8, -1.2, 0.8, -0.8, -1.2, -0.8, -0.8, -1.2, -0.8, 0.8, 1.2, -0.8, -0.8, 1.2, 0.8, -0.8, 1.2)
+points = PoolVector3Array(-0.8, -0.8, -1.2, -0.8, -0.8, 1.2, -0.8, 0.8, -1.2, -0.8, 0.8, 1.2, 0.8, -0.8, -1.2, 0.8, -0.8, 1.2, 0.8, 0.8, -1.2, 0.8, 0.8, 1.2)
 
 [sub_resource id=2 type="ArrayMesh"]
